blocks: make argument values grammar independent
The values for block arguments could only contain either a string or
a numeric value until now. This patch makes it possible to pass
practically anything as block arguments, like white space & other
syntactic elements.

The only constraint is that the value must be enclosed by the parentheses
used to delimit them, e.g.

block destination blk_name(arg(foo bar baz)) {
	`arg`
};

This works both when specifying the default value for an argument, or
when passing it when referencing the macro:

destination d_xxx {
	blk_name(arg(foo bar baz bad));
};

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>

Conflicts:
	lib/cfg-grammar.y