test_buffered_server: This unit test improved to testing correction of bug 33400
In case of calling log_proto_buffered_server_handle_actual_state_version
and the state_version value is 0, then we call the state_handler_get_state
for the new state handler to update the state_version,
but we don't call state_handler_put_state.
If the file doesn't exist, then the transport's fd will be -1,
that means the fstat calling will return -1 in the log_proto_apply_state,
therefore the state_handler_put_state won't be called.

Because of the state_handler_put_state won't be called,
the mapped_counter of the state won't be 0,
therefore the persist_state_grow_store will be blocked.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>