afprog-dest: added program termination on reloadinfo free
The program was not terminated on shutdown of syslog-ng, because at
deinitialization we don't know if the program destination will be
reinitialized, or not.
If the destination won't be reinitialized, the relevant entry in the
GlobalConfig won't be used, therefore its free function will be run.
That is the point, where we can terminate the running program of the
program destination.

Signed-off-by: MÓZES Ádám István <adam.mozes@balabit.com>