logmsg: fix argument list mismatch of log_msg_clone_ack()
OSE commit:
The ack_func prototype was changed when the late acknowledgement mechanism
was introduced, however log_msg_clone_ack was used through an explicit
type cast that suppressed the warning from the compiler. This patch removes
the cast and fixes the argument list.

It fixes github issue #713.

PE notes:
Actually, in OSE, the prototype is fine, but I'd keep this commit
message as-is. The following changes are kept:
 * removing the unnecessary (LMAck) cast
 * make `log_msg_clone_ack` private (aka. static)

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>