hds: the Hierarchical Data Storage is a higher level user of the hashtree
This (hds) will store any monitoring data of syslog-ng.
Every node is able to have a property container,
and every container can store a mount a properties

The Property and the PropertyContainer are interfaces.
The NVPropertyContainer is an implementation of the PropertyContainer interface

The test_hds gives an example how to use the hds.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>