utf8utils: fix two nasty warnings in utf8utils
By accident the _append_escaped_utf8_character() was expecting a pointer
to raw_len in its prototype, instead of the literal length itself, and
then was used as a normal value.

The only reason this worked was the fact that when passing this value
we didn't pass a pointer, but rather an actual integer.

So in turn we were converting an integer to a pointer and then a pointer
to an integer, triggering two warnings but the end actually worked.

This patch fixes this up.

Cherry picked from commit:
github.com/balabit/syslog-ng/commit/efc5280d35e5f8ecbb3c3473dc3612b1f082dec1

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>
Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>