LogReader: avoid to start working in deinitialized state.
On windows file source, it can be possible,
that the deinitialized reader started to work.

Root cause:
===========
AFFileSourceDriver use the same LogReader for every files,
in case of wildcard file sources.
On windows every file source is handled as wildcard file source.
But if the source driver get a NC_FILE_MOVED notification,
then it try to reopen it.
But if it is failed because of any reason,
we deinitialize the LogReader and set our reader to NULL,
triggering that next time when we need a reader (new files
are coming, or other file modified) we create a new reader.
But when we deinitilize the old reader, it is possible,
that it has a registered IO job. If it has registered IO job
it will be scheduled after a while. Reader didn't check
whether it is initialized or not, just start to working.
Deinitilized logpipe has no config object,
therefore the work_perform will crash somewhere.

Solution:
=========
When LogReader starts to execute an IO job (log_reader_work_perform)
it checks, whether it is initialized or not. If it is not initialized
the reader shouldn't start to the real IO job, just skip it.
But If somehow the reader starts to work,
syslog-ng should raise an assert to make easier the investigation.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>