AFFile: make sure, that the notification comes from correct reader
In some special cases the affile can drop a reader and create new one.
Although, the affile set its reader member to NULL,
but the control member of the released reader remains untouched,
and accidentally the released reader can send notification for
the source driver, causing a big problem in file events handling.
In this case we should assert to make easier the investigation of the root cause.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>