logreader: affile shouldn't call the deinit directly in case of file deletion
Root cause
==========

Previous patches solved the problem of starting io job in case of deinitialized reader.
But, there is an other problem. The problem is, what if the io job starts,
and meanwhile the windows filemonitor call the NC_FILE_MOVED notification.
This case causes a race condition. Because a working reader will be deinitialzed.

Solution
========

When a reader should be released by the file source, the file source should only
attempt the deinitialization, not does it directly.
In case of reader's io_job doesn't work, reader can call the deinitialization,
otherwise it will store this request and when the reader finished its job,
it will detect that it should deinitilize itself, and it does the deinitialization

Isn't it racy?
==============

This is not racy, because the working flag of io_job
and pending deinit flag are also set only in main thread.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>