cfg: add global persist-name duplication check
This commit introduces a new GlobalConfig field (persist_names) to make it
possible to avoid persist-name duplications.

Usage example:
"The persist-name of the network based drivers is built from the server address
and port. If there are two network based destination with same port
and server address in a configuration, it means, that their persist name
is the same. It can cause problem for example in case of disk queue usage."

Signed-off-by: László Várady <laszlo.varady@balabit.com>