logsource: fix possible crash when a mangle_callback returns FALSE
Previously, the source's window_size was not decreased when a mangle_callback
returned FALSE, but a slot in the ringbuffer of LateAckTracker was allocated.

Because of the lack of "sync" between the ringbuffer and the window size,
this could cause an assertion failure in late_ack_tracker_track_msg().

This commit fixes the issue by dropping the message when a mangle_callback
returns FALSE.

Fixes #SYSLOGMT-761

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>
Signed-off-by: László Várady <laszlo.varady@balabit.com>