clone_cow: optimization in the log_msg_clone_cow breaks the acknowledgment counting
This optimization try to avoid referencing the clone message,
in case of it has no own payload.
But it is not correct because it breaks the acknowledgment counting.

Why?

Let's see the following situation:
 * There is an original message: O
 * There is clone of O: O'
O' has no own payload, sdata, etc.. (so the cow is not executed on this clone)
We are creating a clone from O': O''. The original of O'' will be the O, not the O'.
This causes the problem, because the acknowledgment will be called for O not for O',
but it should be called for O', because we it waits for this acknowledgment

The result is, that the original message (O) get more acknowledgment,
than it expects causing assertion.

This unit test brings up the issue

Signed-off-by: Viktor Juhasz <viktor.juhasz@balabit.com>