syslog-format: fix portability issues in _normalize_time_cached()
The tm_gmtoff field of the tm structure is not portable (Windows, Solaris,
etc.), so this should not be used as a zone_offset storage.

The value of zone_offset is stored multiple times, because this value may
change after calling the _normalize_time() function.
This is not a real problem because the change only occurs when zone_offset
is -1, but to avoid code coupling, I leave it in this state until the entire
unit has been refactored.

Signed-off-by: László Várady <laszlo.varady@balabit.com>