globals: disable FIPS detection on Windows
FIPS is not supported on Windows, so this check is unnecessary.
Due to the DLL initialization mechanism of Windows, this patch introduced a
memory leak in syslog-ng and syslog-ng Agent.

During the initialization of a dynamic-link library, Windows calls its DllMain
function with the DLL_PROCESS_ATTACH reason code.
After that, when a new thread is created, the system will call DllMain again
from the context of the newly created thread with the DLL_THREAD_ATTACH reason
code. This call allocates memory in libcrypto, but it is not deallocated after
a thread is terminated.

Fixes #SYSLOGDEV-3223

Signed-off-by: László Várady <laszlo.varady@balabit.com>