affile: multiline not active file timeout
Background:

The wildcard file source uses a single logreader/logproto to deal with multiple files.
Because of the single instance it is regularry reloaded with the actual file.
The actual logproto state is saved into the persist storage.

The logreader has a timer for multi line log timeout, but that is always for the actual logreader,
and not for the rest of the files waiting in the persist storage.

Those files that satete are storred in the persist storage does not have active timeout.
The only chance to get the needed lines if the file is updated.

Current patch:

This patch modifies the behavior significantly, because in case of logreader/logproto change,
it does not wait for the actual file for gracefully timeouted but rather forces the buffer flush.
Only after that continues with the logreader replacement.

Signed-off-by: kokan <peter.kokai@balabit.com>