affile: last_msg_parsed should be updated if there was a message
This reverts commit a36dc7a183ab2f2c3215bb2b34677088e3d074b0.
As it turned out the last_msg_time is actually the time of the last successful fetch.
A new last_msg_parsed is introduced, this is updated at the same points as last_msg_time,
the only exception is in the fetch, where it is only updated if logproto could create a
logmsg object.

Signed-off-by: kokan <peter.kokai@balabit.com>