Enforce use-dns-cache(no) when use-dns(no) is set
When using the DNS is disabled, disable the DNS cache too. We do this both in
`cfg_init` (to enforce it at a global level), and in
`log_source_options_init()` (to enforce it at the source level). As there are no
other places where we set these, enforcing at these two places is enough. Even
though `LogWriter` uses the global options, and copies it to its own settings,
it always inherits the global, there is no way to set it on a per-writer or
per-destination basis.

Whenever we enforce the above, we also issue a warning.

Fixes SYSLOGDEV-4151.

Signed-off-by: Gergely Nagy <algernon@balabit.com>