mainloop: sigchld handler chaining
The issue:

The java virtual machine  sets  its own  signal handler. The syslog-ng reaping
happens when the SIGCHLD handler is  called, but the java replaced our SIGCHLD
handler. Therefore every forked process became a zombie, if java is configured.

Solution:

The sigaction is registers the signal handlers, to overcome the java own
registration process in syslog-ng we implement our own sigaction implementation,
which is called instead of the one in libc. Every single case our sigaction
calls the original one, except in case of SIGCHLD. If it is called with SIGCHLD
out of setup phase, the new handler is saved and called, when we are not waiting
for SIGCHLD.

Signed-off-by: Kokan <kokaipeter@gmail.com>