cfg-lexer-include: remove escaping from _glob_pattern_p()
This function is here for compatibility reasons.
Where GLOB_NOMAGIC is not available, this function is used to detect
whether the pattern contains metacharacters.

GLOB_NOMAGIC does not care about escaped characters, so its compat
implementation should behave the same way.


This patch also fixes the @include pragma on Windows when backslashes are
used together with wildcards and the specified directory is empty:

@include 'C:\Program Files\syslog-ng\etc\config\*.conf'

Signed-off-by: László Várady <laszlo.varady@balabit.com>