logreader: do not flush buffers when the reader is started without an fd
log_reader_flush_buffer() is used to flush a partial multiline message
when, for example, the file is moved or deleted (or when the 10seconds
multiline timeout elapsed).

Unfortunately, when the file is moved after syslog-ng has been stopped, it
is not possible to flush the partial message because we have nothing in
our buffers and the old fd (and the file itself) is not available.

Calling log_reader_flush_buffer() in such case produces garbage,
because it tries to extract the partial message from an
uninitialized buffer.

Signed-off-by: László Várady <laszlo.varady@balabit.com>