logreader: fix following file sources
When syslog-ng is stopped while it is reading a file source and the file
source is removed/moved before restarting syslog-ng, it will stop following
the file permanently.

To fix this issue, log_reader_restart_task_handler() should restart
the follow timer even if there is no fd (the file does not exist yet).

Signed-off-by: László Várady <laszlo.varady@balabit.com>