logproto-factor: introduce a stateful flag
This new stateful flag is going to describe if the LogProtoFactory is going to produce a stateless or stateful LogProto.
Normally a LogProto is stateless (prefered), but it is possible to create a LogProto that track the sent messages.
Later the driver could make decisions based on this LogProto attribute.

Signed-off-by: Kokan <kokaipeter@gmail.com>