afinter: add an extra iv_event_post to the internal source
If we failed to load the new configuration during a reload, the
related error messages may stuck in the internal_message_queue
until a new internal message triggers the internal source.
(In most of the cases it is a mark message.) With this manual
event at the end of internal source init, we empty the queue
during the "startup" process.

Signed-off-by: Laszlo Szemere <laszlo.szemere@balabit.com>