mainloop-worker-io: engage vfunction for support symmetry
It looks strange that log_pipe_ref is called outside MainLoopIOWorker
before the job, and it has no log_pipe_unref counterpart in the
neighborhood. To resolve that, the api is changed so that users can
specify their own function that is called before submitting.

With that change, we do not need to call _release when submit early returns.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>