tests: disable test_mainloop
After fixing a config revert worker bug (by adding a iv_task), the fix
indicated that MainLoop needs additional work to be properly testable.

test_mainloop took advantage of the fact that the main loop did not have to
be started in order to test the reload mechanism.
This is not true anymore.

The test can be rewritten after adding a few abstractions to our test lib
that can

- start and stop (schedule a stop event) the main loop, fail after a
specified timeout
- create LogPipes from unit tests, add them to GlobalConfig and initialize
the main loop using this handcrafted configuration
- assert in/on LogPipe methods

All this will probably be implemented only in PE 7.
Until then, I'm disabling the test.

Signed-off-by: László Várady <laszlo.varady@balabit.com>