logproto-buffer: on log-msg-size increase stuck
The LogProtoBufferedServer saves its buffer_size to persist. While this value is
derived from configuration (log-msg-size), thus it do not pick up the value from
configuration. Also the text server splits up messages based on maximum size.
If the log-msg-size is increased (stop/start sylsog-ng), the LogProtoBuffered
server cannot read new bits (as the buffer limits is already reached), but the
text server cannot split the message because the new log-msg-size limit is not
yet reached. Both logic waits for the other to finish.

Signed-off-by: Kokan <kokaipeter@gmail.com>