filemonitor: fix force-directory-polling() on Windows
FileMonitorPoll polls files in a case-sensitive manner, and it can be
configured on Windows too (force-directory-polling).

The Win32 subsystem does not support case-sensitivity for compatibility
reasons, that's why we use case-insensitive filename pattern matching
in the native Windows file monitor.

This should be done in FileMonitorPoll as well.

Note: syslog-ng Agent converts paths to lowercase at startup, so the
force-directory-polling() feature was completely broken when non-lowercase
paths were configured.

Signed-off-by: László Várady <laszlo.varady@balabit.com>